Worked VA Djarvis

+++ processes.py
@@ -3,19 +3,31 @@ import os
 import subprocess
 import wmi
 
+from speaker import va_speak
 
 PRIORITY_PROGRAMS = ['Calculator.exe']
 
 f = wmi.WMI()
 
+def check_process():
+    try:
+        asyncio.run(processes())
+        va_speak("Вывожу все работающие программы")
+    except:
+        print('Error')
 
-async def processes():
+
+def processes_slow():
     for process in f.Win32_Process():
         if process.Name in PRIORITY_PROGRAMS:
             print(f'{process.ProcessID:<10} {process.Name}')
 
 
-def start_browser(close_program):
+def processes(close_program=None):
+    """
+
+    :param close_program: if you want to close some write it's 'name.exe'
+    """
     programs_in_work = []
     try:
         _procces = subprocess.check_output("tasklist /fo csv /nh")
@@ -31,33 +43,25 @@ def start_browser(close_program):
 
                 programs_in_work.append([element, element_pid])
 
-                if element == close_program.lower():
-                    print(f"Закрываю {element_pid} {element}\n")
 
-                    subprocess.call(f"taskkill /pid {element_pid}")
+                if not close_program is None:
+                    if element == close_program.lower():
+                        print(f"Закрываю {element_pid} {element}\n")
+
+                        subprocess.call(f"taskkill /pid {element_pid}")
+
+        if close_program is None:
+            return programs_in_work
 
     except exec() as er:
 
         print('Error: ', er)
 
 
-# async def processes():
-#     pr = []
-#     for process in f.Win32_Process():
-#         pr.append(process.Name)
-#     pr.sort()
-#     for el in pr:
-#         print(el)
-#         if el in PRIORITY_PROGRAMS['programs']:
-#             print(f'######## {el} ########')
-
 if __name__ == '__main__':
-    # asyncio.run(processes())
-    # print(start_browser)
-    # for my_computer in f:
-    #     print(my_computer)
-    prog = input()
-    start_browser(prog)
+
+    # prog = input()
+    print(processes())
